OCE-57: NPE in DirectInventoryDatasource (#69)
Fixed an NPE that would be triggered if the same alarm was deleted more than once. You could cause this simply by sending a delete for the same alarm back to back. In practice I suspect this was caused by a delete coming in explicitly followed by a delete derived from a processing a snapshot.

I also simplified the thread-safety by synchronizing all the handle* methods as I think there is very little cost in doing so and it seems safer overall (I couldn't cause any threading issues by stress testing it, but better to be safer here I think).