NMS-15391: clean up dependencies on core and sentinel
To be compatible with Meridian, dependencies have been adjusted
to be more in-line with how things work across Horizon/Meridian.

On the Meridian side, the `meridian` packages use `Provides:`
to specify availability of `opennms-core` or `opennms-sentinel`.
On the ALEC side, we should only care about OPA version, and
that _some_ version of core or sentinel are installed. OPA
compatibility numbers should be the determining factor for
specific version compat.